plan: move test.
